birdy/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.5.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
namex==0.0.8
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
pycparser==2.22
Pygments==2.18.0
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sqlparse==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
